requirements.txt
absl-py==1.0.0
aioredis==1.3.1
amqp==5.1.1
anyio==3.6.1
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
auditlog3==1.0.1
autobahn==22.5.1
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
basemap==1.3.4
basemap-data==1.3.2
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
blinker==1.4
boto3==1.24.2
botocore==1.27.2
Brlapi==0.7.0
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.0.0
celery==5.2.7
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==3.4.7
cssselect2==0.4.1
cupshelpers==1.0
cycler==0.11.0
daphne==3.0.2
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.4
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.43.0
django-bootstrap-breadcrumbs==0.9.2
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-ckeditor==6.0.0
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
django-dynamic-breadcrumbs==0.4.2
django-environ==0.4.5
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.9.0
django-js-asset==2.0.0
django-smtp-ssl==1.0
django-storages==1.12.3
django-taggit==3.0.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.0.0
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.0.12
Fiona==1.8.21
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fonttools==4.25.2
future==0.18.2
gast==0.5.3
GDAL==3.0.4
GeoAlchemy2==0.12.1
geographiclib==1.52
geojson==2.5.0
geopandas==0.11.1
geopy==2.2.0
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
gpg===1.13.1-unknown
greenlet==1.1.2
grpcio==1.46.0
gunicorn==20.1.0
h11==0.12.0
h3==3.7.4
h5py==3.6.0
hiredis==2.0.0
html5lib==1.1
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
incremental==21.3.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kazam==1.4.5
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.2
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.1
lockfile==0.12.2
louis==3.12.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.7
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.4
munch==2.5.0
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
nltk==3.7
notebook==6.4.11
numpy==1.22.3
oauthlib==3.1.1
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.10
opt-einsum==3.3.0
OWSLib==0.19.1
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
paytmchecksum==1.7.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.5.2
plotly==4.4.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
psycopg2==2.8.6
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pycups==1.9.73
pydyf==0.0.3
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyperclip==1.8.2
pyphen==0.11.0
pyproj==3.3.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.1
pyshp==2.3.0
PySimpleSOAP==1.16.2
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.4
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
redis==4.3.4
regex==2022.4.24
reportlab==3.5.34
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
rfc3986==1.5.0
rsa==4.8
s3transfer==0.6.0
SecretStorage==2.3.1
Send2Trash==1.8.0
sendgrid==6.4.7
service-identity==21.1.0
Shapely==1.8.2
simplejson==3.16.0
sip==4.19.21
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.1
stack-data==0.2.0
starkbank-ecdsa==2.0.3
stripe==3.2.0
suds==1.1.2
suds-jurko==0.6
suds-py3==1.4.5.0
systemd-python==234
tablib==3.2.1
tekore==4.4.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
terminado==0.13.3
textblob==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
timezonefinder==6.0.1
tinycss2==1.1.0
torbrowser-launcher==0.3.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
twilio==6.48.0
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
tzdata==2022.5
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.6
usb-creator==0.3.7
utm==0.7.0
vine==5.0.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
weasyprint==53.0
webencodings==0.5.1
Werkzeug==2.1.2
whitenoise==5.3.0
wrapt==1.14.1
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zeep==4.1.0
zipp==3.8.0
zope.interface==5.4.0
zopfli==0.1.8
